Διόρθωση του createpatient και δημιουργία του editpatient

+++ DoctorAppointments/Repository/PatientRepository.cs
@@ -4,7 +4,8 @@ namespace DoctorAppointments.Repository
 {
     public class PatientRepository : IPatientRepository
     {
-        private DoctorAppointmentsDbContext _context;
+        private readonly DoctorAppointmentsDbContext _context;
+
         public PatientRepository(DoctorAppointmentsDbContext context)
         {
             _context = context ?? throw new ArgumentNullException(nameof(context));
@@ -12,6 +13,8 @@ namespace DoctorAppointments.Repository
 
         public void Add(Patient patient)
         {
+            if (patient == null) throw new ArgumentNullException(nameof(patient));
+
             _context.Patients.Add(patient);
             _context.SaveChanges();
         }
@@ -19,30 +22,33 @@ namespace DoctorAppointments.Repository
         public void Delete(int id)
         {
             var patient = GetById(id);
+            if (patient == null) throw new ArgumentNullException("Patient not found.");
+
             _context.Patients.Remove(patient);
             _context.SaveChanges();
         }
 
         public IEnumerable<Patient> GetAll()
         {
-            var patients = _context.Patients.ToList();
-            return patients;
+            return _context.Patients.ToList();
         }
 
         public Patient? GetByAMKA(string amka)
         {
-            var patient = _context.Patients.FirstOrDefault(x => x.AMKA == amka);
-            return patient;
+            if (string.IsNullOrWhiteSpace(amka)) return null;
+
+            return _context.Patients.FirstOrDefault(x => x.AMKA == amka);
         }
 
         public Patient? GetById(int id)
         {
-            var patient = _context.Patients.Find(id);
-            return patient;
+            return _context.Patients.Find(id);
         }
 
         public void Update(Patient patient)
         {
+            if (patient == null) throw new ArgumentNullException(nameof(patient));
+
             _context.Patients.Update(patient);
             _context.SaveChanges();
         }
